+++ packages/plugin-compiler/src/test/kotlin/io/realm/kotlin/compiler/GenerationExtensionTest.kt
@@ -34,6 +34,7 @@ import io.realm.kotlin.internal.schema.PropertyMetadata
 import io.realm.kotlin.internal.schema.SchemaMetadata
 import io.realm.kotlin.types.BaseRealmObject
 import io.realm.kotlin.types.RealmObject
+import io.realm.kotlin.types.TypedRealmObject
 import org.junit.Test
 import java.io.File
 import kotlin.reflect.KClass
@@ -371,15 +372,14 @@ class GenerationExtensionTest {
                         get() = TODO("Not yet implemented")
                     override val properties: List<PropertyMetadata>
                         get() = TODO("Not yet implemented")
+                    override val clazz: KClass<out TypedRealmObject>?
+                        get() = TODO("Not yet implemented")
                     override val className: String
                         get() = TODO("Not yet implemented")
                     override val primaryKeyProperty: PropertyMetadata?
                         get() = TODO("Not yet implemented")
                     override val isEmbeddedRealmObject: Boolean
                         get() = TODO("Not yet implemented")
-                    override val clazz: KClass<out BaseRealmObject>?
-                        get() = TODO("Not yet implemented")
-
                     override fun get(propertyKey: PropertyKey): PropertyMetadata? {
                         TODO("Not yet implemented")
                     }
@@ -390,6 +390,7 @@ class GenerationExtensionTest {
                         TODO("Not yet implemented")
                     }
                 }
+
                 override fun get(classKey: ClassKey): ClassMetadata? {
                     TODO("Not yet implemented")
                 }